Move the implementation of http services to be under server package
